Clean up recruitment and create forms * Remove non-model code from models. This is now in utils. * Create forms app for easy access to CustomField and ExtraField outside of recruitment.
