+++ ais/common/settings.py
@@ -40,6 +40,7 @@ INSTALLED_APPS = (
     'api',
     'news',
     'crispy_forms',
+    'forms',
 )
 
 MIDDLEWARE_CLASSES = (
